shoppingManage
Version 1.1: If you used arrays to store user information, product information, and shopping history, upgrade to ArrayList, LinkedList, etc. (this requirement was completed in version 1.0); In addition, this version supports user information, product information, shopping history information stored in a text file.

PS :(a) do not need in the administrator/user in the operation process, at any time to read and write text files, the operation process, the program read and write is the collection of framework class objects. You simply read user information, product information, and shopping history from the ArrayList object and write them to a text file when the program exits, and then read them from the text file the next time the program starts. That is, write a text file (there may be multiple files) once upon exit and read a text file once upon startup. (b) The user password cannot be directly stored in plain text. To use MD5 encryption processing before storage, the corresponding password authentication function should also be upgraded. (This feature is not implemented yet, I'm still testing it, so I haven't added it to the program yet)